refactor : 게임 재시작 로직 분리
