+++ src/service/gameService.js
@@ -1,0 +1,10 @@
+import { YES } from "../constants/constants.js";
+import { lottoController } from "../controller/lottoController.js";
+
+export const gameService =  {
+    reStart(input){
+        if(input===YES){
+            lottoController.run();
+        } 
+    },
+}
